# swagger_rcs_gateway.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "RCS Gateway API with Authentication",
...
        "contact": {
            "name": "Suporte RCS Gateway",
            "email": "suporte@exemplo.com"
        },
        "version": "2.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autentica\u00e7\u00e3o e Administra\u00e7\u00e3o"
                ],
                "summary": "Login",
                "description": "Login de usu\u00e1rio",
                "operationId": "login_api_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/clients": {
            "post": {
                "tags": [
                    "Autentica\u00e7\u00e3o e Administra\u00e7\u00e3o"
                ],
                "summary": "Create Client",
                "description": "Criar novo cliente",
                "operationId": "create_client_api_admin_clients_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Autentica\u00e7\u00e3o e Administra\u00e7\u00e3o"
                ],
                "summary": "List Clients",
                "description": "Listar clientes",
                "operationId": "list_clients_api_admin_clients_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    },
                                    "title": "Response List Clients Api Admin Clients Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/client/send-message": {
            "post": {
                "tags": [
                    "Servi\u00e7os para Clientes"
                ],
                "summary": "Send Message",
                "description": "Endpoint principal para clientes enviarem mensagens RCS\n\nEste endpoint permite que clientes autenticados enviem mensagens RCS\nespecificando o tipo (basic ou single) e opcionalmente um template.\n\n**Par\u00e2metros:**\n- **client_code**: C\u00f3digo do cliente (deve corresponder ao cliente autenticado)\n- **user_code**: C\u00f3digo do usu\u00e1rio (opcional)\n- **message_type**: Tipo da mensagem (\"basic\" ou \"single\")\n- **template_id**: ID do template da plataforma (opcional)\n- **phone_numbers**: Lista de n\u00fameros de telefone\n- **variables**: Vari\u00e1veis para substitui\u00e7\u00e3o no template\n- **content**: Conte\u00fado personalizado (se n\u00e3o usar template)\n- **callback_url**: URL de callback (opcional, usa padr\u00e3o do cliente)\n- **campaign_name**: Nome da campanha (opcional)\n\n**Exemplo de uso:**\n```json\n{\n    \"client_code\": \"CLI_ABC123\",\n    \"user_code\": \"user001\",\n    \"message_type\": \"basic\",\n    \"template_id\": \"template_001\",\n    \"phone_numbers\": [\"5511999999999\"],\n    \"variables\": {\n        \"nome\": \"Jo\u00e3o\",\n        \"produto\": \"Smartphone\"\n    },\n    \"callback_url\": \"https://meusite.com/callback\",\n    \"campaign_name\": \"Promocao_Janeiro\"\n}\n```",
                "operationId": "send_message_api_client_send_message_post",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "API Key do cliente",
                            "title": "X-Api-Key"
                        },
                        "description": "API Key do cliente"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientSendMessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientSendMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/client/messages": {
            "get": {
                "tags": [
                    "Servi\u00e7os para Clientes"
                ],
                "summary": "Get Messages",
                "description": "Consultar mensagens enviadas pelo cliente\n\nEste endpoint permite que clientes consultem suas mensagens enviadas\ncom filtros opcionais por data, status, n\u00famero de telefone, etc.\n\n**Par\u00e2metros de consulta:**\n- **client_code**: C\u00f3digo do cliente\n- **start_date**: Data inicial (formato: YYYY-MM-DD)\n- **end_date**: Data final (formato: YYYY-MM-DD)\n- **status**: Status da mensagem (pending, sent, delivered, read, failed)\n- **phone_number**: N\u00famero de telefone espec\u00edfico\n- **limit**: Limite de resultados (padr\u00e3o: 100)\n- **offset**: Offset para pagina\u00e7\u00e3o (padr\u00e3o: 0)\n\n**Exemplo de resposta:**\n```json\n[\n    {\n        \"id\": \"123\",\n        \"client_message_id\": \"CLI_ABC123_a1b2c3d4\",\n        \"phone_number\": \"5511999999999\",\n        \"status\": \"delivered\",\n        \"message_type\": \"basic\",\n        \"campaign_name\": \"Promocao_Janeiro\",\n        \"sent_at\": \"2025-01-15T10:30:00Z\",\n        \"delivered_at\": \"2025-01-15T10:31:00Z\",\n        \"created_at\": \"2025-01-15T10:30:00Z\"\n    }\n]\n```",
                "operationId": "get_messages_api_client_messages_get",
                "parameters": [
                    {
                        "name": "client_code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Client Code"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Start Date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "End Date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Status"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Phone Number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "API Key do cliente",
                            "title": "X-Api-Key"
                        },
                        "description": "API Key do cliente"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClientMessageResponse"
                                    },
                                    "title": "Response Get Messages Api Client Messages Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/rcs/basic": {
            "post": {
                "tags": [
                    "RCS Original (Compatibilidade)"
                ],
                "summary": "Send Basic Message Legacy",
                "description": "[LEGACY] Envio de mensagem RCS Basic (mantido para compatibilidade)\n\n\u26a0\ufe0f **Deprecated**: Use `/api/client/send-message` para novos projetos",
                "operationId": "send_basic_message_legacy_api_rcs_basic_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RCSBasicRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rcs/single": {
            "post": {
                "tags": [
                    "RCS Original (Compatibilidade)"
                ],
                "summary": "Send Single Message Legacy",
                "description": "[LEGACY] Envio de mensagem RCS Single (mantido para compatibilidade)\n\n\u26a0\ufe0f **Deprecated**: Use `/api/client/send-message` para novos projetos",
                "operationId": "send_single_message_legacy_api_rcs_single_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RCSSingleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/messages": {
            "get": {
                "tags": [
                    "RCS Original (Compatibilidade)"
                ],
                "summary": "List Messages Legacy",
                "description": "[LEGACY] Listar mensagens (mantido para compatibilidade)\n\n\u26a0\ufe0f **Deprecated**: Use `/api/client/messages` para novos projetos",
                "operationId": "list_messages_legacy_api_messages_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rcs/callback": {
            "post": {
                "tags": [
                    "RCS Original (Compatibilidade)"
                ],
                "summary": "Receive Callback Legacy",
                "description": "[LEGACY] Receber callback da API RCS (mantido para compatibilidade)\n\nEste endpoint ainda \u00e9 usado pela Eugen para enviar callbacks.",
                "operationId": "receive_callback_legacy_api_rcs_callback_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "title": "Callback Data"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClientCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "rcs_account": {
                        "type": "string",
                        "title": "Rcs Account"
                    },
                    "api_token": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Api Token"
                    },
                    "allowed_message_types": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Allowed Message Types",
                        "default": [
                            "basic",
                            "single"
                        ]
                    },
                    "max_messages_per_day": {
                        "type": "string",
                        "title": "Max Messages Per Day",
                        "default": "1000"
                    },
                    "callback_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Callback Url"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "email",
                    "rcs_account"
                ],
                "title": "ClientCreate"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LoginRequest": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password"
                ],
                "title": "LoginRequest"
            },
...
            "Token": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "title": "Access Token"
                    },
                    "token_type": {
                        "type": "string",
                        "title": "Token Type",
                        "default": "bearer"
                    },
                    "expires_in": {
                        "type": "integer",
                        "title": "Expires In"
                    },
                    "client_code": {
                        "type": "string",
                        "title": "Client Code"
                    }
                },
                "type": "object",
                "required": [
                    "access_token",
                    "expires_in",
                    "client_code"
                ],
                "title": "Token"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}